Suggestion development (#36)
Made few changes to suggestion feature.
